Remove TorDisposableBase.
